add detail page to events
Also fixed some unit tests that broke because of the new constraint on the email column

+++ partEZ/database/migrations/2016_02_01_141154_add_status_to_invites.php
@@ -27,9 +27,12 @@ class AddStatusToInvites extends Migration
      */
     public function down()
     {
-        Schema::table('invites', function($table)
+        if (Schema::hasColumn('invites', 'status'))
         {
-            $table->dropColumn('status');
-        });
+            Schema::table('invites', function($table)
+            {
+                $table->dropColumn('status');
+            });
+        }
     }
 }
